# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Unified Service Scheduler API","description":"Resource-constrained appointment booking for dealership service departments.","contact":{"name":"Keyloop Scheduler"},"license":{"name":"Apache-2.0"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Appointments","description":"Book and query service appointments"},{"name":"Reference data","description":"Lookup dealerships and service types"}],"paths":{"/api/v1/appointments":{"post":{"tags":["Appointments"],"summary":"Book a service appointment","description":"Reserves a qualified technician and a service bay for the full service duration.","operationId":"book","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppointmentResponse"}}}}}}},"/api/v1/service-types":{"get":{"tags":["Reference data"],"summary":"List service types","operationId":"serviceTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ServiceTypeView"}}}}}}}},"/api/v1/dealerships":{"get":{"tags":["Reference data"],"summary":"List dealerships","operationId":"dealerships","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DealershipView"}}}}}}}},"/api/v1/appointments/{id}":{"get":{"tags":["Appointments"],"summary":"Get an appointment by id","operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AppointmentResponse"}}}}}}},"/api/v1/appointments/availability":{"get":{"tags":["Appointments"],"summary":"Check availability","description":"Non-binding probe for a proposed window; does not reserve resources.","operationId":"availability","parameters":[{"name":"dealershipId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"serviceTypeId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"desiredStart","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AvailabilityResponse"}}}}}}}},"components":{"schemas":{"BookingRequest":{"required":["dealershipId","desiredStart","serviceTypeId","vehicleId"],"type":"object","properties":{"dealershipId":{"type":"integer","format":"int64"},"vehicleId":{"type":"integer","format":"int64"},"serviceTypeId":{"type":"integer","format":"int64"},"desiredStart":{"type":"string","format":"date-time"}}},"AppointmentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"status":{"type":"string"},"dealership":{"$ref":"#/components/schemas/Ref"},"customer":{"$ref":"#/components/schemas/Ref"},"vehicle":{"$ref":"#/components/schemas/VehicleRef"},"technician":{"$ref":"#/components/schemas/Ref"},"serviceBay":{"$ref":"#/components/schemas/Ref"},"serviceType":{"$ref":"#/components/schemas/ServiceTypeRef"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"}}},"Ref":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ServiceTypeRef":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"}}},"VehicleRef":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"vin":{"type":"string"},"make":{"type":"string"},"model":{"type":"string"}}},"ServiceTypeView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"},"requiredSkill":{"type":"string"}}},"DealershipView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"openingTime":{"$ref":"#/components/schemas/LocalTime"},"closingTime":{"$ref":"#/components/schemas/LocalTime"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"AvailabilityResponse":{"type":"object","properties":{"available":{"type":"boolean"},"start":{"type":"string","format":"date-time"},"end":{"type":"string","format":"date-time"},"availableTechnicians":{"type":"integer","format":"int32"},"availableBays":{"type":"integer","format":"int32"},"reason":{"type":"string"}}}}}}